pyproject.toml
[project]
name = "skin-cancer-classification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastai==2.7.17",
    "fastapi==0.115.0",
    "httpx==0.28.0",
    "joblib==1.4.2",
    "keras==3.6.0",
    "matplotlib==3.9.0",
    "opencv-python>=4.11.0.86",
    "pandas==2.2.0",
    "pillow==11.0.0",
    "plotly==5.24.0",
    "psutil==5.9.0",
    "pydantic==2.11.0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.16",
    "requests==2.32.0",
    "scikit-learn==1.5.0",
    "seaborn==0.13.0",
    "streamlit==1.40.0",
    "tensorflow==2.17.0",
    "typing-extensions==4.14.0",
    "uvicorn[standard]==0.32.0",
]
